+++ src/interact/vite/pagesProviderModule.d.ts
@@ -1,11 +1,13 @@
 declare module 'interact:page-modules' {
     import {Page} from "@combostrap/interact/types";
 
+    type pageModule = { module: Page, path: string };
+
     export function getModulePage(opts: {
         path: string;
         notFoundPath?: string;
-    }): Page | undefined;
+    }): pageModule | undefined;
 
-    export const modulePages: Record<string, Page>;
+    export const modulePages: Record<string, pageModule>;
     export default getModulePage;
 }